Move report to utils.
